Support for future VOS Docker Images
Added support for using new and future VOS Docker images, which do not
include the 'vos-interruptible.sh' wrapper. With these changes,
'vos-interruptible.sh' is no longer used. Instead, docker-compose is
configured to automatically send SIGINT to Virtuoso instead of the
default SIGTERM, and waits a maximum of 2 minutes for a graceful stop
(normal shutdown) instead of the default 10 seconds.